console: Refactor repetitions
